refactor(services): split cell/spawner.rs into submodules
Split into spawner/{missions,dialogs,npcs,respawners,stargates,regions,abilities,loot}.rs by feature, with tests in spawner/tests.rs and re-exports in mod.rs.

Public API unchanged. No behavior changes.
