+++ crates/services/src/cell/spawner/dialogs.rs
@@ -1,0 +1,43 @@
+//! Dialog set map cache.
+//!
+//! Maps `dialog_set_map_id → (dialog_id, interaction_flags)` for use by
+//! `add_dialog_set` content actions at runtime.
+
+use sqlx::PgPool;
+
+/// Cached row from `resources.dialog_set_maps`, used by `add_dialog_set` content actions.
+#[derive(Debug, Clone)]
+pub struct DialogSetMapEntry {
+    pub dialog_id: i32,
+    pub interaction_flags: i64,
+}
+
+/// Load the `dialog_set_maps` lookup table from the database.
+///
+/// Maps `dialog_set_map_id → (dialog_id, interaction_flags)` so that
+/// `add_dialog_set` actions can resolve at runtime without per-action DB queries.
+pub async fn load_dialog_set_maps(
+    pool: &PgPool,
+) -> Result<std::collections::HashMap<i32, DialogSetMapEntry>, sqlx::Error> {
+    use sqlx::Row;
+
+    let rows = sqlx::query(
+        "SELECT dialog_set_map_id, dialog_id, interaction_flags \
+         FROM resources.dialog_set_maps"
+    )
+    .fetch_all(pool)
+    .await?;
+
+    let mut map = std::collections::HashMap::with_capacity(rows.len());
+    for r in &rows {
+        let id: i32 = r.get("dialog_set_map_id");
+        let dialog_id: Option<i32> = r.get("dialog_id");
+        let interaction_flags: i64 = r.get("interaction_flags");
+        if let Some(dialog_id) = dialog_id {
+            map.insert(id, DialogSetMapEntry { dialog_id, interaction_flags });
+        }
+    }
+
+    tracing::info!(count = map.len(), "Loaded dialog_set_maps cache");
+    Ok(map)
+}
